+++ src/Delizious.Ini/IniDocument.cs
@@ -774,5 +774,74 @@
 
             this.iniDocument.WriteComment(sectionName, comment, mode);
         }
+
+        /// <summary>
+        /// <para>
+        /// Writes the comment for the property according to the given mode.
+        /// </para>
+        /// <para>
+        /// When mode is <see cref="CommentWriteMode.Fail"/> and the section does not exist,
+        /// throws a <see cref="SectionNotFoundException"/> 
+        /// </para>
+        /// <para>
+        /// When mode is <see cref="CommentWriteMode.Fail"/> and the property does not exist,
+        /// throws a <see cref="PropertyNotFoundException"/> 
+        /// </para>
+        /// <para>
+        /// When <paramref name="mode"/> is <see cref="CommentWriteMode.Ignore"/>,
+        /// it is silently ignored if the section or property does not exist.
+        /// </para>
+        /// </summary>
+        /// <param name="sectionName">
+        /// The name of the section containing the property.
+        /// </param>
+        /// <param name="propertyKey">
+        /// The key of the property to write the comment.
+        /// </param>
+        /// <param name="comment">
+        /// The comment to write for the property.
+        /// </param>
+        /// <param name="mode">
+        /// The mode that specifies how to write the comment.
+        /// </param>
+        /// <exception cref="ArgumentNullException">
+        /// <para><paramref name="sectionName"/> is <c>null</c>.</para>
+        /// <para>- or -</para>
+        /// <para><paramref name="propertyKey"/> is <c>null</c>.</para>
+        /// <para>- or -</para>
+        /// <para><paramref name="comment"/> is <c>null</c>.</para>
+        /// <para>- or -</para>
+        /// <para><paramref name="mode"/> is <c>null</c>.</para>
+        /// </exception>
+        /// <exception cref="SectionNotFoundException">
+        /// <paramref name="mode"/> is <see cref="CommentWriteMode.Fail"/> and the specified section does not exist.
+        /// </exception>
+        /// <exception cref="PropertyNotFoundException">
+        /// <paramref name="mode"/> is <see cref="CommentWriteMode.Fail"/> and the specified property does not exist.
+        /// </exception>
+        public void WriteComment(SectionName sectionName, PropertyKey propertyKey, Comment comment, CommentWriteMode mode)
+        {
+            if (sectionName is null)
+            {
+                throw new ArgumentNullException(nameof(sectionName));
+            }
+
+            if (propertyKey is null)
+            {
+                throw new ArgumentNullException(nameof(propertyKey));
+            }
+
+            if (comment is null)
+            {
+                throw new ArgumentNullException(nameof(comment));
+            }
+
+            if (mode is null)
+            {
+                throw new ArgumentNullException(nameof(mode));
+            }
+
+            this.iniDocument.WriteComment(sectionName, propertyKey, comment, mode);
+        }
     }
 }