+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.16.0] - 2025-04-06
+### Added
+- Enable writing the comment for a property of an INI document with given section name, property key and comment write mode ([#168](https://github.com/oliverzick/Delizious-Ini/issues/168))
+
 ## [1.15.0] - 2025-03-30
 ### Added
-- Enable writing the comment of a section from INI document for given section name and comment write mode ([#160](https://github.com/oliverzick/Delizious-Ini/issues/160))
+- Enable writing the comment for a section of an INI document with given section name and comment write mode ([#160](https://github.com/oliverzick/Delizious-Ini/issues/160))
 
 ## [1.14.0] - 2025-03-15
 ### Added